benchmark/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "toke-benchmark"
version = "0.1.0"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "M. Karwalski"}]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pyyaml>=6",
    "ruff>=0.4",
    "mypy>=1.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
strict = true
python_version = "3.11"
